requirements.txt
TestLink-API-Python-client
requests
xlrd
pytest
pytest-allure-adaptor